wip(UIKIT-1016,features, ProductsSwitcher): Исправления по замечаниям ПР
