+++ packages/features/src/productSwitcher/components/IdentityProductSwitcher/IdentityProductSwitcher.tsx
@@ -1,18 +1,12 @@
-import { MouseEvent, useContext, useMemo, useState } from 'react';
 import {
-  ConfigContext,
   ContentState,
   IconButton,
   MenuGroup,
   Typography,
-  WidgetProduct,
   useMenu,
 } from '@astral/ui';
 import { ErrorFillSm, ProductsFillMd } from '@astral/icons';
 
-import { IdentityTenantsDTO } from '../../types';
-
-import { getIdentityTenants, getTenantsProducts } from './utils';
 import {
   Logo,
   ProductItem,
@@ -21,6 +15,7 @@ import {
   TitleErrorContainer,
   WidgetMenu,
 } from './styles';
+import { useLogic } from './hooks';
 
 export const ASTRAL_IDENTITY_DEFAULT_TENANT = 'astral';
 
@@ -34,61 +29,17 @@ export type IdentityProductSwitcherType = {
 export const IdentityProductSwitcher = ({
   identityUrl,
 }: IdentityProductSwitcherType) => {
-  const [tenantsProductsMap, setTenantsProductsMap] =
-    useState<Record<string, WidgetProduct[]>>();
-
-  const [tenants, setTenants] = useState<IdentityTenantsDTO[]>();
-  const [tenantId, setTenantId] = useState<string>(
-    ASTRAL_IDENTITY_DEFAULT_TENANT,
-  );
-
-  const [isLoading, setIsLoading] = useState(false);
-  const [isError, setIsError] = useState(false);
-  const { captureException } = useContext(ConfigContext);
   const { open, anchorRef, handleOpenMenu, handleCloseMenu } = useMenu();
 
-  const productList = useMemo(
-    () => tenantsProductsMap?.[tenantId],
-    [tenantsProductsMap, tenantId],
-  );
-
-  const handleShowProducts = async () => {
-    setIsError(false);
-    handleOpenMenu();
-
-    if (!tenantsProductsMap) {
-      setIsLoading(true);
-
-      try {
-        const tenantsInfo = await getIdentityTenants(identityUrl);
-        const tenantsIds = tenantsInfo.map(({ id }) => id);
-        const tenantsProducts = await getTenantsProducts(
-          identityUrl,
-          tenantsIds,
-        );
-
-        if (tenantsIds[0] !== tenantId) {
-          setTenantId(tenantsIds[0]);
-        }
-
-        setTenants(tenantsInfo);
-        setTenantsProductsMap(tenantsProducts);
-      } catch (error) {
-        setIsError(true);
-        captureException(error);
-      } finally {
-        setIsLoading(false);
-      }
-    }
-  };
-  const handleChangeTenant = async (
-    _event: MouseEvent<HTMLElement>,
-    selectedValue: string,
-  ) => {
-    if (selectedValue) {
-      setTenantId(selectedValue);
-    }
-  };
+  const {
+    isLoading,
+    isError,
+    tenantId,
+    tenants,
+    productList,
+    handleChangeTenant,
+    handleShowProducts,
+  } = useLogic(identityUrl, handleOpenMenu);
 
   return (
     <>
