+++ packages/features/src/productSwitcher/components/IdentityProductSwitcher/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useLogic } from './useLogic';
