+++ packages/features/src/productSwitcher/components/IdentityProductSwitcher/hooks/useLogic/index.ts
@@ -1,0 +1,1 @@
+export * from './useLogic';
